requirements.txt
Django==1.8
django-haystack==2.6.0
elasticsearch>=5.0.0,<6.0.0
